+++ CHANGELOG.md
@@ -1,7 +1,21 @@
 ﻿
 # Changelog
 
-## Latest Version : v0.1.0
+## Latest Version : v0.1.1
+
+## [v0.1.1] - 2025-07-19
+### Added
+
+- CorrelationId Middleware has been added.
+- CorrelationId Middleware extension has been added.
+- CorrelationId Delegate Handler has been added.
+- PaymentOrchestration service extensions seperated.
+- Serilog has been configured.
+
+### Changed
+
+### Fixed
+
 
 ## [v0.1.0] - 2025-07-18
 ### Added